+++ src/functional/otf/compilation/build_systems/cmake.py
@@ -6,9 +6,8 @@ import subprocess
 from typing import Optional
 
 from functional.otf import languages, stages
-from functional.otf.compilation import build_data, common, compiler
+from functional.otf.compilation import build_data, cache, common, compiler
 from functional.otf.compilation.build_systems import cmake_lists
-from functional.program_processors.builders import cache
 
 
 @dataclasses.dataclass
