+++ src/functional/program_processors/runners/gtfn_cpu.py
@@ -21,10 +21,9 @@ import numpy as np
 from functional.iterator import ir as itir
 from functional.otf import languages, stages, workflow
 from functional.otf.binding import pybind
-from functional.otf.compilation import compiler
+from functional.otf.compilation import cache, compiler
 from functional.otf.compilation.build_systems import compiledb
 from functional.program_processors import processor_interface as fpi
-from functional.program_processors.builders import cache
 from functional.program_processors.codegens.gtfn import gtfn_module
 from functional.program_processors.source_modules import cpp_gen
