+++ tests/functional_tests/fencil_processor_tests/builder_tests/test_importer.py
@@ -15,7 +15,7 @@
 
 import textwrap
 
-from functional.program_processors.builders import importer
+from functional.otf.compilation import importer
 
 
 def test_import_from_path(tmp_path):
